Updated some commands
Here we have updated some of the commands. We have also added more commands and are expecting update V4 to roll out soon. How-ever as now there seems to be an issue with the skills/buy system where the buy system would NOT post a message if a skill is missing.